refactoring, debugging
