+++ src/mcr/causality/scms/examples.py
